+++ .github/workflows/ci.yml
@@ -1,0 +1,55 @@
+name: CI
+
+# Запускаем этот workflow на каждый push, pull_request в ветки dev и main
+on:
+  push:
+    branches: [dev, main]
+  pull_request:
+    branches: [dev]
+
+jobs:
+  lint-and-test-and-build:
+    runs-on: ubuntu-latest
+
+    steps:
+      # 1) Загрузить код из репозитория
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      # 2) Установить Node.js (LTS-версия 18.x)
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: '18'
+          cache: 'npm'
+
+      # 3) Установить зависимости
+      - name: Install dependencies
+        run: npm ci
+
+      # 4) Запустить ESLint
+      - name: Lint code
+        run: npm run lint
+
+      # 5) Запустить тесты (unit и интеграционные, если есть)
+      - name: Run tests
+        run: npm run test
+
+      # 6) Запустить сборку проекта
+      - name: Build project
+        run: npm run build
+
+      # (опционально) 7) Проверить на уязвимости npm-пакетов
+      - name: Audit npm dependencies
+        run: npm audit --audit-level=moderate
+
+      # (опционально) 8) Генерация coverage-отчёта
+      - name: Run coverage
+        run: npm run test:coverage
+        continue-on-error: true
+
+      # (опционально) 9) Загрузить coverage на Codecov
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v3
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }} # поставить токен в Settings > Secrets
