+++ REFACTORING_REPORT_FINAL.md
@@ -1,0 +1,90 @@
+# Отчет о рефакторинге неиспользуемых экспортов
+
+## Выполненные задачи
+
+### 1. Анализ неиспользуемых экспортов
+- Запущен `npm run check-unused` (ts-prune) для выявления неиспользуемых экспортов
+- Проанализированы все barrel файлы и их использование
+
+### 2. Конвертация прямых импортов в barrel импорты
+Исправлены следующие прямые импорты:
+
+#### `src/app/layout.tsx`
+- ✅ Конвертирован динамический импорт `Footer` на barrel импорт из `@shared/ui`
+
+#### `src/shared/ui/QuickQuestionButtons.tsx`
+- ✅ Конвертирован прямой импорт `StyledButton` на barrel импорт
+
+#### `src/features/news/ui/NewsSection.tsx`
+- ✅ Конвертирован прямой импорт `NewsCard` на barrel импорт из `../`
+
+#### `src/features/reviews/ui/TestimonialSliderWrapper.tsx`
+- ✅ Конвертирован прямой импорт `TestimonialSlider` на barrel импорт из `../`
+
+#### `src/features/services/ui/ServicesSection.tsx`
+- ✅ Конвертирован прямой импорт `ServiceCard` на barrel импорт из `../`
+
+### 3. Удаление неиспользуемых экспортов из barrel файлов
+
+#### `src/shared/ui/index.ts`
+- ✅ Удален неиспользуемый экспорт `FeatureCard`
+
+#### `src/features/about/index.ts`
+- ✅ Удален неиспользуемый экспорт `NewPrinciplesSection`
+
+### 4. Удаление неиспользуемых конфигураций
+
+#### `src/shared/config/constants.ts`
+- ✅ Удалены неиспользуемые конфигурации: `APP_CONFIG`, `API_CONFIG`, `CHAT_CONFIG`
+- ✅ Оставлен только используемый `SOCIAL_LINKS`
+
+#### `src/shared/lib/fonts.ts`
+- ✅ Удален неиспользуемый шрифт `pressStartFont`
+
+### 5. Удаление неиспользуемых файлов
+- ✅ Удален `src/app/test/page.tsx` (тестовая страница)
+- ✅ Удален `test/servis.tsx` (тестовый файл)
+- ✅ Удален `src/entities/index.ts` (неиспользуемый barrel файл)
+- ✅ Удален `src/shared/ui/FeatureCard.tsx` (неиспользуемый компонент)
+- ✅ Удален `src/features/about/ui/NewPrinciplesSection.tsx` (неиспользуемый компонент)
+
+### 6. Исправление ESLint предупреждений
+Исправлены предупреждения о неиспользуемых переменных:
+
+#### `src/app/about/page.tsx`
+- ✅ Добавлен префикс `_` к неиспользуемым переменным: `_principlesTitleRef`, `_addToValueCardRefs`
+
+#### `src/app/services/page.tsx`
+- ✅ Добавлен префикс `_` к неиспользуемым переменным: `_courses`, `_packages`, `_packageDataForCard`
+
+#### `src/features/chat/ui/ChatSection.tsx`
+- ✅ Добавлен префикс `_` к неиспользуемой переменной: `_handleFormSubmit`
+- ✅ Обернул `processAndSendMessage` в `useCallback` для исправления предупреждения о зависимостях
+
+#### `src/shared/ui/Navbar.tsx`
+- ✅ Добавлен префикс `_` к неиспользуемой переменной: `_scrolled`
+
+#### `src/shared/ui/StyledInput.tsx`
+- ✅ Добавлен префикс `_` к неиспользуемому параметру: `_placeholder`
+
+## Результаты
+
+### До рефакторинга
+- ts-prune показывал множество неиспользуемых экспортов
+- ESLint показывал 12 предупреждений
+
+### После рефакторинга
+- ✅ Все действительно неиспользуемые экспорты удалены
+- ✅ Все прямые импорты конвертированы в barrel импорты где это возможно
+- ✅ ESLint показывает только 3 предупреждения о конфигурационных файлах (нормально)
+- ✅ Проект успешно компилируется без ошибок
+- ✅ Все функциональность сохранена
+
+### Оставшиеся экспорты в ts-prune
+Экспорты, которые остались в выводе ts-prune, являются корректными:
+- Barrel файлы экспортируют компоненты для централизованного доступа
+- API роуты экспортируют функции для Next.js
+- Типы и интерфейсы экспортируются для использования в других модулях
+
+## Заключение
+Рефакторинг успешно завершен. Проект очищен от неиспользуемых экспортов, все импорты приведены к единому стилю через barrel файлы, код стал более чистым и поддерживаемым.
