--- src/entities/index.ts
@@ -1,2 +1,0 @@
-export * from './news';
-export * from './newsData';
