+++ src/features/about/index.ts
@@ -1,4 +1,3 @@
-export { default as NewPrinciplesSection } from './ui/NewPrinciplesSection';
 export { BookCards } from './ui/BookCards';
 export { default as FlipCard } from './ui/FlipCard';
 export { default as FlipCardStyles } from './ui/FlipCardStyles';
